+++ db/migrate/20180806133214_add_amount_types_to_expense_services.rb
@@ -1,11 +1,11 @@
 class AddAmountTypesToExpenseServices < ActiveRecord::Migration[5.1]
   def change
-    add_column :expense_services, :max_cost, :money
-    add_column :expense_services, :min_cost, :money
-    add_column :expense_services, :mid_cost, :money
-    add_column :expense_services, :max_amount, :integer
-    add_column :expense_services, :mid_amount, :integer
-    add_column :expense_services, :min_amount, :integer
-    add_column :expense_services, :amount, :integer
+    add_column :expense_services, :max_cost, :float
+    add_column :expense_services, :min_cost, :float
+    add_column :expense_services, :mid_cost, :float
+    add_column :expense_services, :max_amount, :float
+    add_column :expense_services, :mid_amount, :float
+    add_column :expense_services, :min_amount, :float
+    add_column :expense_services, :amount, :float
   end
 end
